+++ packages/client/tsup.config.ts
@@ -2,7 +2,10 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
   tsconfig: "./tsconfig.json",
-  entry: ["src/index.ts"],
+  entry: {
+    index: "src/index.ts",
+    utils: "src/utils/index.ts",
+  },
   format: ["esm", "cjs"],
   outDir: "dist",
   bundle: true,
@@ -15,10 +18,5 @@ export default defineConfig({
   skipNodeModulesBundle: true,
   onSuccess: async () => {
     console.log("Build completed successfully!");
-  },
-  outExtension({ format }) {
-    return {
-      js: format === "cjs" ? ".cjs" : ".mjs",
-    };
   },
 });